# [Quant] Vectorize scalar remainder in quantized kernel for normalization