--- a/aten/src/ATen/cpu/vec/vec256/vec256_qint.h
+++ b/aten/src/ATen/cpu/vec/vec256/vec256_qint.h
@@ -597,8 +597,20 @@ struct Vectorized<c10::quint8> : public Vectorizedqi {
         }
     }
 
-    static Vectorized<c10::quint8> loadu(const void* ptr) {
-        return Vectorized<c10::quint8>(ptr);
+    static Vectorized<c10::quint8> loadu(const void* ptr, int64_t count = size()) {
+        if (count == size()) {
+          return Vectorized<c10::quint8>(ptr);
+        }
+        __at_align__ value_type tmp_values[size()];
+        // Ensure uninitialized memory does not change the output value See https://github.com/pytorch/pytorch/issues/32502
+        // for more details. We do not initialize arrays to zero using "={0}" because gcc would compile it to two
+        // instructions while a loop would be compiled to one instruction.
+        for (const auto i : c10::irange(size())) {
+          tmp_values[i] = 0;
+        }
+        std::memcpy(
+            tmp_values, reinterpret_cast<const value_type*>(ptr), count * sizeof(value_type));
+        return _mm256_loadu_si256((const __m256i*)tmp_values);
     }
 
  private: